// src/app/models/typeColors.ts
export enum TypeColors {
  Normal = "#A7A777",
  Fighting = "#C03028",
  Flying = "#A790F0",
  Poison = "#9F409F",
  Ground = "#E0C068",
  Rock = "#B89F38",
  Bug = "#A7B820",
  Ghost = "#705898",
  Steel = "#B8B8D0",
  Fire = "#F08030",
  Water = "#6890F0",
  Grass = "#77C750",
  Electric = "#F8D030",
  Psychic = "#F85888",
  Ice = "#98D8D8",
  Dragon = "#7038F8",
  Dark = "#705848",
  Fairy = "#EE99AB"
}

export const types = [
  "Normal",
  "Fighting",
  "Flying",
  "Poison",
  "Ground",
  "Rock",
  "Bug",
  "Ghost",
  "Steel",
  "Fire",
  "Water",
  "Grass",
  "Electric",
  "Psychic",
  "Ice",
  "Dragon",
  "Dark",
  "Fairy"
];